feat: standardize limit and offset handling across files
- Add metadata import `metadata.New(map[string]string{"total": strconv.Itoa(total)})` to `restaurant_grpc_server.go`
- Change `Limit` and `Offset` types to `int64` in `restaurant.go`
- Add `metadata` import to `restaurant_grpc_server.go`
- Set default `Limit` to `100` in `mongodb.go`
- Change `Limit` and `Offset` handling in `mongodb.go`

+++ entity/domain/restaurant/repo/restaurant.go
@@ -9,8 +9,8 @@ import (
 
 // ListCondition is a struct that defines the condition for listing restaurants
 type ListCondition struct {
-	Limit  int
-	Offset int
+	Limit  int64
+	Offset int64
 }
 
 // IRestaurantRepo is an interface that defines the methods that the restaurant repository should implement
